notes from .ai
