+++ tests/modular_pipelines/krea2/test_modular_pipeline_krea2_turbo.py
@@ -16,7 +16,13 @@
 
 from diffusers.modular_pipelines import Krea2TurboAutoBlocks, Krea2TurboModularPipeline
 
-from ..test_modular_pipelines_common import ModularPipelineTesterMixin
+from ..testing_utils import (
+    BaseModularPipelineTesterConfig,
+    ModularLoadingTesterMixin,
+    ModularMemoryTesterMixin,
+    ModularPipelineTesterMixin,
+    ModularWorkflowTesterMixin,
+)
 
 
 KREA2_TURBO_WORKFLOWS = {
@@ -32,11 +38,10 @@ KREA2_TURBO_WORKFLOWS = {
 }
 
 
-class TestKrea2TurboModularPipelineFast(ModularPipelineTesterMixin):
+class Krea2TurboModularPipelineTesterConfig(BaseModularPipelineTesterConfig):
     pipeline_class = Krea2TurboModularPipeline
     pipeline_blocks_class = Krea2TurboAutoBlocks
     pretrained_model_name_or_path = "hf-internal-testing/tiny-krea2-turbo-modular-pipe"
-
     params = frozenset(["prompt", "height", "width"])
     batch_params = frozenset(["prompt"])
     expected_workflow_blocks = KREA2_TURBO_WORKFLOWS
@@ -54,5 +59,19 @@ class TestKrea2TurboModularPipelineFast(ModularPipelineTesterMixin):
         }
         return inputs
 
+
+class TestKrea2TurboModularPipelineFast(Krea2TurboModularPipelineTesterConfig, ModularPipelineTesterMixin):
     def test_inference_batch_single_identical(self):
         super().test_inference_batch_single_identical(expected_max_diff=5e-3)
+
+
+class TestKrea2TurboModularPipelineLoading(Krea2TurboModularPipelineTesterConfig, ModularLoadingTesterMixin):
+    pass
+
+
+class TestKrea2TurboModularPipelineWorkflow(Krea2TurboModularPipelineTesterConfig, ModularWorkflowTesterMixin):
+    pass
+
+
+class TestKrea2TurboModularPipelineMemory(Krea2TurboModularPipelineTesterConfig, ModularMemoryTesterMixin):
+    pass
